% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette_accessors.R
\name{get_hbma_cell_type_palette}
\alias{get_hbma_cell_type_palette}
\title{Get HBMA cell type color palette}
\usage{
get_hbma_cell_type_palette()
}
\value{
A named character vector
}
\description{
Returns a named vector of HBMA colors by cell type.
}
