% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette_scales.R
\name{hbma_cell_type_color_scale}
\alias{hbma_cell_type_color_scale}
\title{ggplot2 color scale for cell types using HBMA convention}
\usage{
hbma_cell_type_color_scale(name = "Cell type", ...)
}
\arguments{
\item{name}{Legend title (default = "Cell type")}

\item{...}{Additional arguments passed to \code{scale_color_manual()}}
}
\value{
A ggplot2 color scale
}
\description{
ggplot2 color scale for cell types using HBMA convention
}
\examples{
library(ggplot2)
set.seed(2464)
df <- data.frame(
  cell_type = c(
    "immune", "vascular", "cortex GABAergic", "astrocyte", "ependymal", "OPC",
    "oligodendrocyte", "striatum MSN", "striatum non-MSN",
    "cortex glutamatergic"
  ), stringsAsFactors = FALSE
)
df$value <- runif(nrow(df))

ggplot(df, aes(x = cell_type, y = value, color = cell_type)) +
  geom_point() +
  hbma_cell_type_color_scale()

}
