% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbma_cell_type_palette}
\alias{hbma_cell_type_palette}
\title{HBMA cell type palette}
\format{
A named character vector
}
\usage{
hbma_cell_type_palette
}
\description{
A named character vector mapping cell type names to hex colors.
Typically accessed via \code{\link[=get_hbma_cell_type_palette]{get_hbma_cell_type_palette()}}.
}
\keyword{datasets}
